+++ app/src/main/java/com/fieldbook/tracker/activities/DataGridActivity.kt
@@ -8,6 +8,7 @@ import android.util.Log
 import android.util.TypedValue
 import android.view.Menu
 import android.view.MenuItem
+import android.view.View
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
@@ -52,6 +53,7 @@ import javax.inject.Inject
 import androidx.core.content.edit
 import androidx.databinding.DataBindingUtil
 import com.fieldbook.tracker.databinding.ActivityDataGridBinding
+import com.fieldbook.tracker.utilities.InsetHandler
 import com.google.firebase.crashlytics.FirebaseCrashlytics
 import eu.wewox.lazytable.LazyTableState
 import eu.wewox.lazytable.lazyTablePinConfiguration
@@ -120,6 +122,8 @@ class DataGridActivity : ThemedActivity(), CoroutineScope by MainScope() {
 
         setSupportActionBar(binding.toolbar)
 
+        InsetHandler.setupStandardInsets(binding.root, binding.toolbar)
+
         if (supportActionBar != null) {
             supportActionBar?.title = null
             supportActionBar?.setDisplayHomeAsUpEnabled(true)
@@ -149,6 +153,8 @@ class DataGridActivity : ThemedActivity(), CoroutineScope by MainScope() {
                 }
             }
         }
+
+        onBackPressedDispatcher.addCallback(this, standardBackCallback())
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {