+++ app/src/main/java/com/fieldbook/tracker/activities/FieldEditorActivity.kt
@@ -32,7 +32,6 @@ import com.fieldbook.tracker.async.ImportRunnableTask
 import com.fieldbook.tracker.brapi.BrapiInfoDialogFragment
 import com.fieldbook.tracker.database.DataHelper
 import com.fieldbook.tracker.database.models.ObservationUnitModel
-import com.fieldbook.tracker.dialogs.FieldCreatorDialogFragment
 import com.fieldbook.tracker.dialogs.FieldSortDialogFragment
 import com.fieldbook.tracker.dialogs.ListAddDialog
 import com.fieldbook.tracker.dialogs.ListSortDialog
@@ -143,6 +142,18 @@ class FieldEditorActivity : BaseFieldActivity(), FieldSortController {
             }
         }
 
+    private val fieldCreatorLauncher =
+        registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
+            if (result.resultCode == RESULT_OK) {
+                val fieldId = result.data?.getIntExtra("fieldId", -1) ?: -1
+                if (fieldId != -1) {
+                    fieldSwitcher.switchField(fieldId)
+                    queryAndLoadFields()
+                    startFieldDetailFragment(fieldId)
+                }
+            }
+        }
+
     override fun getLayoutResourceId(): Int = R.layout.activity_fields
 
     override fun getToolbarId(): Int = R.id.field_toolbar
@@ -186,18 +197,7 @@ class FieldEditorActivity : BaseFieldActivity(), FieldSortController {
         mAdapter = FieldAdapter(this, this, fieldGroupController, false)
         mAdapter.setOnFieldActionListener(object : FieldAdapter.OnFieldActionListener {
             override fun onFieldDetailSelected(fieldId: Int) {
-                val fragment = FieldDetailFragment()
-                val args = Bundle()
-                args.putInt("fieldId", fieldId)
-                fragment.arguments = args
-
-                // Disable touch events on the RecyclerView
-                recyclerView.isEnabled = false
-
-                supportFragmentManager.beginTransaction()
-                    .replace(android.R.id.content, fragment, "FieldDetailFragmentTag")
-                    .addToBackStack(null)
-                    .commit()
+                startFieldDetailFragment(fieldId)
             }
 
             override fun onFieldSetActive(fieldId: Int) {
@@ -492,15 +492,7 @@ class FieldEditorActivity : BaseFieldActivity(), FieldSortController {
                 0 -> if (checkDirectory()) loadLocalPermission()
                 1 -> if (checkDirectory()) loadCloud()
                 2 -> {
-                    val dialog = FieldCreatorDialogFragment(this)
-                    dialog.fieldCreationCallback =
-                        object : FieldCreatorDialogFragment.FieldCreationCallback {
-                            override fun onFieldCreated(studyDbId: Int) {
-                                fieldSwitcher.switchField(studyDbId)
-                                queryAndLoadFields()
-                            }
-                        }
-                    dialog.show(supportFragmentManager, "FieldCreatorDialogFragment")
+                    fieldCreatorLauncher.launch(FieldCreatorActivity.getIntent(this))
                 }
 
                 3 -> loadBrAPI()