+++ app/src/main/java/com/fieldbook/tracker/activities/TraitEditorActivity.java
@@ -31,6 +31,7 @@ import android.widget.LinearLayout.LayoutParams;
 import android.widget.ListView;
 import android.widget.Toast;
 
+import androidx.activity.OnBackPressedCallback;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.widget.Toolbar;
@@ -59,6 +60,7 @@ import com.fieldbook.tracker.preferences.GeneralKeys;
 import com.fieldbook.tracker.preferences.PreferenceKeys;
 import com.fieldbook.tracker.utilities.CSVWriter;
 import com.fieldbook.tracker.utilities.FileUtil;
+import com.fieldbook.tracker.utilities.InsetHandler;
 import com.fieldbook.tracker.utilities.SharedPreferenceUtils;
 import com.fieldbook.tracker.utilities.TapTargetUtil;
 import com.fieldbook.tracker.utilities.Utils;
@@ -315,6 +317,7 @@ public class TraitEditorActivity extends ThemedActivity implements TraitAdapterC
         super.onCreate(savedInstanceState);
 
         setContentView(R.layout.activity_traits);
+        setupTraitEditorInsets();
 
         Toolbar toolbar = findViewById(R.id.toolbar);
 
@@ -345,6 +348,8 @@ public class TraitEditorActivity extends ThemedActivity implements TraitAdapterC
 
         FloatingActionButton fab = findViewById(R.id.newTrait);
         fab.setOnClickListener(v -> showImportDialog());
+
+        setupBackCallback();
     }
 
     // Implement the interface
@@ -766,12 +771,23 @@ public class TraitEditorActivity extends ThemedActivity implements TraitAdapterC
         builder.setTitle(getString(R.string.traits_toolbar_delete_all));
         builder.setMessage(getString(R.string.dialog_delete_traits_message));
 
-        builder.setPositiveButton(getString(android.R.string.yes), onPositive);
+        builder.setPositiveButton(getString(R.string.dialog_delete), onPositive);
         builder.setNegativeButton(getString(R.string.dialog_no), onNegative);
         builder.setOnDismissListener(onDismiss);
 
         AlertDialog alert = builder.create();
         alert.show();
+    }
+
+    private void setupBackCallback() {
+        OnBackPressedCallback backCallback = new OnBackPressedCallback(true) {
+            @Override
+            public void handleOnBackPressed() {
+                CollectActivity.reloadData = true;
+                finish();
+            }
+        };
+        getOnBackPressedDispatcher().addCallback(this, backCallback);
     }
 
     @Override
@@ -953,5 +969,13 @@ public class TraitEditorActivity extends ThemedActivity implements TraitAdapterC
         }
         queryAndLoadTraits();
         refreshTraitDetailFragment();
+    }
+
+    private void setupTraitEditorInsets() {
+        View rootView = findViewById(android.R.id.content);
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        FloatingActionButton fab = findViewById(R.id.newTrait);
+
+        InsetHandler.INSTANCE.setupStandardInsets(rootView, toolbar);
     }
 }