+++ app/src/main/java/com/fieldbook/tracker/activities/brapi/io/BrapiTraitImporterActivity.kt
@@ -6,8 +6,7 @@ import android.os.Bundle
 import android.view.MenuItem
 import android.view.View
 import android.widget.ProgressBar
-import android.widget.Toast
-import androidx.activity.result.contract.ActivityResultContracts
+import androidx.activity.OnBackPressedDispatcher
 import androidx.appcompat.widget.Toolbar
 import com.fieldbook.tracker.R
 import com.fieldbook.tracker.activities.ThemedActivity
@@ -28,6 +27,7 @@ import kotlinx.coroutines.MainScope
 import kotlinx.coroutines.cancel
 import javax.inject.Inject
 import androidx.core.content.edit
+import com.fieldbook.tracker.utilities.InsetHandler
 
 @AndroidEntryPoint
 class BrapiTraitImporterActivity : BrapiTraitImportAdapter.TraitLoader, ThemedActivity(),
@@ -70,12 +70,16 @@ class BrapiTraitImporterActivity : BrapiTraitImportAdapter.TraitLoader, ThemedAc
 
         parseIntentExtras()
 
+        val rootView = findViewById<View>(android.R.id.content)
+        InsetHandler.setupStandardInsets(rootView, toolbar)
+
+        onBackPressedDispatcher.addCallback(this, standardBackCallback())
     }
 
     private fun parseIntentExtras() {
         val dbIds = intent.getStringArrayListExtra(EXTRA_TRAIT_DB_ID)
         if (dbIds == null) {
-            setResult(Activity.RESULT_CANCELED)
+            setResult(RESULT_CANCELED)
             finish()
         }
 
@@ -124,7 +128,7 @@ class BrapiTraitImporterActivity : BrapiTraitImportAdapter.TraitLoader, ThemedAc
 
             prefs.edit { remove(BrapiTraitFilterActivity.FILTER_NAME) }
 
-            setResult(Activity.RESULT_OK)
+            setResult(RESULT_OK)
             finish()
         }
     }
@@ -147,8 +151,8 @@ class BrapiTraitImporterActivity : BrapiTraitImportAdapter.TraitLoader, ThemedAc
 
         when (item.itemId) {
             android.R.id.home -> {
-                setResult(Activity.RESULT_CANCELED)
-                finish()
+                setResult(RESULT_CANCELED)
+                onBackPressedDispatcher.onBackPressed()
                 return true
             }
         }