+++ app/src/main/java/com/fieldbook/tracker/adapters/ImageAdapter.kt
@@ -8,6 +8,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageButton
 import android.widget.ImageView
+import android.widget.TextView
 import androidx.camera.view.PreviewView
 import androidx.cardview.widget.CardView
 import androidx.constraintlayout.widget.ConstraintLayout
@@ -68,6 +69,7 @@ class ImageAdapter(private val context: Context, private val listener: ImageItem
         val cardView: CardView = view.findViewById(R.id.list_item_image_cv)
         val imageView: ImageView = view.findViewById(R.id.list_item_image_iv)
         val closeButton: ImageButton = view.findViewById(R.id.list_item_image_close_btn)
+        val labelView: TextView = view.findViewById<TextView>(R.id.list_item_image_label_tv)
 
         init {
             // Define click listener for the ViewHolder's View.
@@ -89,6 +91,8 @@ class ImageAdapter(private val context: Context, private val listener: ImageItem
 
             itemView.tag = model
 
+            labelView.visibility = View.GONE
+
             try {
 
                 val (actualWidth, actualHeight) =
@@ -103,10 +107,16 @@ class ImageAdapter(private val context: Context, private val listener: ImageItem
                     height = actualHeight
                 }
 
-                val preview = if (model.uri == "NA") {
+                val preview = if (model.uri?.contains("content://") != true) {
+                    if (model.uri != "NA") {
+                        labelView.text = model.uri
+                        labelView.visibility = View.VISIBLE
+                    }
                     val data = context.resources.assets.open("na_placeholder.jpg").readBytes()
                     BitmapFactory.decodeByteArray(data, 0, data.size)
-                } else BitmapLoader.getPreview(view.context, model.uri, model.orientation)
+                } else {
+                    BitmapLoader.getPreview(view.context, model.uri, model.orientation)
+                }
 
                 imageView.setImageBitmap(preview)