+++ app/src/main/java/com/fieldbook/tracker/database/Migrator.kt
@@ -14,6 +14,7 @@ import com.fieldbook.tracker.database.migrators.GroupMigratorVersion14
 import com.fieldbook.tracker.database.migrators.MulticatToCategoricalVersion20
 import com.fieldbook.tracker.database.migrators.SpectralMigratorVersion16
 import com.fieldbook.tracker.database.migrators.TraitAliasSynonymVersion18
+import com.fieldbook.tracker.database.migrators.StudyConfigurationVersion17
 import com.fieldbook.tracker.objects.TraitObject
 
 /**
@@ -413,6 +414,18 @@ class Migrator {
                     Log.e(TAG, "Failed to migrate to version 16", it)
                 }
                 .onSuccess {
+                    Log.d(TAG, "Migrated to version 16")
+                }
+        }
+
+        fun migrateToVersion17(db: SQLiteDatabase) {
+
+            StudyConfigurationVersion17().migrate(db)
+                .onFailure {
+                    Log.e(TAG, "Failed to migrate to version 17", it)
+                }
+                .onSuccess {
+                    Log.d(TAG, "Migrated to version 17")
                     Log.d(TAG, "Migrated to version 16")
                 }
         }