+++ app/src/main/java/com/fieldbook/tracker/database/dao/StudyDao.kt
@@ -222,6 +222,9 @@ class StudyDao {
             it.searchAttribute = this["observation_unit_search_attribute"]?.toString()
             it.groupId = this["group_id"]?.toString()?.toIntOrNull()
             it.archived = this["is_archived"].toString() == "true"
+            it.startCorner = this["start_corner"]?.toString()
+            it.walkingDirection = this["walking_direction"]?.toString()
+            it.walkingPattern = this["walking_pattern"]?.toString()
         }
 
         fun getAllFieldObjects(sortOrder: String): ArrayList<FieldObject> = withDatabase { db ->
@@ -280,6 +283,9 @@ class StudyDao {
                     observation_unit_search_attribute,
                     is_archived,
                     group_id,
+                    start_corner,
+                    walking_direction,
+                    walking_pattern,
                     (SELECT COUNT(DISTINCT observation_unit_attribute_name) FROM observation_units_attributes AS A
                         JOIN observation_units_values AS V 
                             ON V.observation_unit_attribute_db_id = A.internal_id_observation_unit_attribute
@@ -330,10 +336,7 @@ class StudyDao {
 
                 val cursor = db.rawQuery("""
                     SELECT ov.observation_variable_name, ov.observation_variable_field_book_format, COUNT(*) as count, GROUP_CONCAT(o.value, '|') as observations,
-                    (SELECT COUNT(DISTINCT observation_unit_id) 
-                        FROM observations 
-                        JOIN observation_variables AS ov ON ov.${ObservationVariable.PK} = observations.${ObservationVariable.FK}
-                        WHERE study_id = ?) AS distinct_obs_units,
+                    COUNT(DISTINCT observation_unit_id) AS distinct_obs_units,
                     (SELECT COUNT(*) FROM observation_units WHERE study_id = ?) AS total_obs_units,
                     (SELECT v.observation_variable_attribute_value 
                      FROM observation_variable_values v
@@ -344,7 +347,7 @@ class StudyDao {
                     WHERE o.study_id = ? AND o.observation_variable_db_id > 0
                     GROUP BY ov.observation_variable_name, ov.observation_variable_field_book_format
                     ORDER BY ov.${if (sortOrder == "visible") "position" else sortOrder} COLLATE NOCASE ASC
-                """, arrayOf(studyId.toString(), studyId.toString(), studyId.toString()))
+                """, arrayOf(studyId.toString(), studyId.toString()))
 
                 if (cursor.moveToFirst()) {
                     do {
@@ -413,6 +416,9 @@ class StudyDao {
                     put("count", field.entryCount)
                     put("observation_levels", field.observationLevel)
                     put("trial_name", field.trialName)
+                    put("start_corner", field.startCorner)
+                    put("walking_direction", field.walkingDirection)
+                    put("walking_pattern", field.walkingPattern)
                 }).toInt()
 
                 try {