+++ app/src/main/java/com/fieldbook/tracker/traits/formats/parameters/Parameters.kt
@@ -3,7 +3,7 @@ package com.fieldbook.tracker.traits.formats.parameters
 enum class Parameters {
 
     FORMAT, NAME, DEFAULT_VALUE, DETAILS, MAXIMUM, MINIMUM, CATEGORIES, CAMERA, CLOSE_KEYBOARD,
-    CROP_IMAGE, USE_DAY_OF_YEAR, DISPLAY_VALUE, RESOURCE_FILE,
+    CROP_IMAGE, SAVE_IMAGE, USE_DAY_OF_YEAR, DISPLAY_VALUE, RESOURCE_FILE,
     DECIMAL_PLACES,
     MATHEMATICAL_SYMBOLS,
     MULTIPLE_CATEGORIES,
@@ -20,9 +20,6 @@ enum class Parameters {
             "category",
             "closeKeyboardOnOpen",
             "cropImage",
-            "useDayOfYear",
-            "displayValue",
-            "resourceFile"
         )
     }
 }