+++ app/src/main/java/com/fieldbook/tracker/views/TraitBoxView.kt
@@ -119,6 +119,8 @@ class TraitBoxView : ConstraintLayout {
         } else {
             traitsStatusBarRv?.visibility = GONE
         }
+
+        updateTraitBoxArrows()
     }
 
     fun handleTraitTypeWrapping() {
@@ -187,7 +189,14 @@ class TraitBoxView : ConstraintLayout {
         }
 
         updateTraitsStatusBar()
+        updateTraitBoxArrows()
+    }
 
+    private fun updateTraitBoxArrows() { // hide arrows if only one trait is active
+        val shouldShowArrows = visibleTraitsList.size > 1
+
+        traitLeft.visibility = if (shouldShowArrows) VISIBLE else GONE
+        traitRight.visibility = if (shouldShowArrows) VISIBLE else GONE
     }
 
     fun getRecyclerView(): RecyclerView? {